Refactor all public facing pages to a single controller
